backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.39
psycopg[binary]==3.2.6
alembic==1.15.2
pydantic==2.10.6
pydantic-settings==2.8.1
httpx==0.28.1
sympy==1.13.3
python-multipart==0.0.20
pillow==11.2.1
pytesseract==0.3.13
python-dotenv==1.0.1
pytest==8.3.5
